+++ src/kernel/src/dataframe.h
@@ -25,8 +25,17 @@
 
 namespace vita
 {
-/// The type used as class id in classification tasks.
+/// The type used as class ID in classification tasks.
 using class_t = std::size_t;
+
+/// Used for access protection.
+///
+/// \ see https://stackoverflow.com/q/3220009/3235496
+class dataframe_only
+{
+  friend class dataframe;
+  dataframe_only() {}
+};
 
 ///
 /// A 2-dimensional labeled data structure with columns of potentially
@@ -38,15 +47,55 @@ using class_t = std::size_t;
 /// - is modelled on the corresponding *pandas* object;
 /// - is a forward iterable collection of "monomorphic" examples (all samples
 ///   have the same type and arity);
-/// - accepts many different kinds of input: CSV and XRFF
-///   (http://weka.wikispaces.com/XRFF) files.
+/// - accepts many different kinds of input: CSV and XRFF files.
 ///
 class dataframe
 {
 public:
   // ---- Structures ----
   struct example;
-  struct column;
+  struct params;
+
+  /// Information about the collection of columns (type, name, output index).
+  class columns_info
+  {
+  public:
+    /// Information about a single column of the dataset.
+    struct column_info
+    {
+      std::string       name =                 {};
+      category_t category_id = undefined_category;
+    };
+
+    using size_type = std::size_t;
+
+    columns_info();
+
+    const column_info &operator[](size_type i) const { return cols_[i]; }
+    column_info &operator[](size_type i) { return cols_[i]; }
+
+    size_type size() const { return cols_.size(); }
+    bool empty() const { return cols_.empty(); }
+
+    auto begin() const { return cols_.begin(); }
+    auto begin() { return cols_.begin(); }
+    auto end() const { return cols_.end(); }
+    auto end() { return cols_.end(); }
+
+    auto back() const { return cols_.back(); }
+    auto back() { return cols_.back(); }
+
+    void pop_back() { cols_.pop_back(); }
+    void push_back(const column_info &);
+    void push_front(const column_info &);
+
+    template<class T> void build(const T &, bool, category_set &);
+
+    void swap_category(category_t, category_t, dataframe_only);
+
+  private:
+    std::vector<column_info> cols_;
+  };
 
   // ---- Aliases ----
   using examples_t = std::vector<example>;
@@ -63,8 +112,10 @@ public:
 
   // ---- Constructors ----
   dataframe();
-  explicit dataframe(std::istream &, filter_hook_t = nullptr);
-  explicit dataframe(const std::filesystem::path &, filter_hook_t = nullptr);
+  explicit dataframe(std::istream &);
+  dataframe(std::istream &, const params &);
+  explicit dataframe(const std::filesystem::path &);
+  dataframe(const std::filesystem::path &, const params &);
 
   // ---- Iterators ----
   using iterator = typename examples_t::iterator;
@@ -76,21 +127,25 @@ public:
   iterator end();
   const_iterator end() const;
 
+  value_type front() const;
+  value_type &front();
+
   // ---- Modifiers ----
   void clear();
   iterator erase(iterator, iterator);
 
   // ---- Convenience ----
-  std::size_t read(const std::filesystem::path &, filter_hook_t = nullptr);
-  std::size_t read_csv(std::istream &, filter_hook_t = nullptr);
-  std::size_t read_xrff(std::istream &, filter_hook_t = nullptr);
+  std::size_t read(const std::filesystem::path &);
+  std::size_t read(const std::filesystem::path &, const params &);
+  std::size_t read_csv(std::istream &);
+  std::size_t read_csv(std::istream &, const params &);
+  std::size_t read_xrff(std::istream &);
+  std::size_t read_xrff(std::istream &, const params &);
   bool operator!() const;
 
   void push_back(const example &);
 
   const category_set &categories() const;
-
-  const std::vector<column> &columns() const;
 
   std::size_t size() const;
   bool empty() const;
@@ -102,15 +157,17 @@ public:
 
   bool debug() const;
 
+  columns_info columns;
+
 private:
   bool read_record(const record_t &);
-  example to_example(const record_t &, bool, bool);
+  example to_example(const record_t &, bool);
 
   class_t encode(const std::string &);
 
-  std::size_t read_csv(const std::filesystem::path &, filter_hook_t);
-  std::size_t read_xrff(const std::filesystem::path &, filter_hook_t);
-  std::size_t read_xrff(tinyxml2::XMLDocument &, filter_hook_t);
+  std::size_t read_csv(const std::filesystem::path &, const params &);
+  std::size_t read_xrff(const std::filesystem::path &, const params &);
+  std::size_t read_xrff(tinyxml2::XMLDocument &, const params &);
 
   void swap_category(category_t, category_t);
 
@@ -118,12 +175,6 @@ private:
   // input strings are converted into integers by this map and the `encode`
   // static function.
   std::map<std::string, class_t> classes_map_;
-
-  // How is the dataset organized? Sometimes we have a dataset header (XRFF
-  // file format), other times it has to be implicitly derived (e.g. CSV).
-  // `header_[0]` is the output column (it contains informations about
-  // problem's output).
-  std::vector<column> header_;
 
   // What are the categories we're dealing with?
   // Note: `category_[0]` is the output category.
@@ -178,6 +229,22 @@ inline class_t label(const dataframe::example &e)
   return std::get<D_INT>(e.output);
 }
 
+struct dataframe::params
+{
+  //params() : has_header(false), filter(nullptr), output_index(0) {}
+
+  /// `true` when the CSV file has a header.
+  /// \remark Used only when reading CSV files.
+  bool has_header = false;
+
+  /// A filter and transform function applied when reading data.
+  filter_hook_t filter = nullptr;
+
+  /// Index of the column containing the output value (label).
+  /// \remark Used only when reading CSV files.
+  std::optional<std::size_t> output_index = 0;
+};
+
 ///
 /// Get the output value for a given example.
 ///
@@ -189,25 +256,8 @@ inline class_t label(const dataframe::example &e)
 template<class T>
 T label_as(const dataframe::example &e)
 {
-  const auto &v(e.output);
-
-  if (std::holds_alternative<D_DOUBLE>(v))
-    return static_cast<T>(std::get<D_DOUBLE>(v));
-
-  if (std::holds_alternative<D_INT>(v))
-    return static_cast<T>(std::get<D_INT>(v));
-
-  return 0;
+  return lexical_cast<T>(e.output);
 }
-
-///
-/// Informations about a "column" of the dataset.
-///
-struct dataframe::column
-{
-  std::string       name;
-  category_t category_id;
-};
 
 }  // namespace vita