+++ src/test/tests.cc
@@ -36,3 +36,4 @@
 #include "test/symbol_set.cc"
 #include "test/team.cc"
 #include "test/terminal.cc"
+#include "test/value.cc"