+++ src/utility/utility.cc
@@ -165,6 +165,20 @@ double lexical_cast<double>(const vita::value_t &v)
   }
 }
 
+template<>
+int lexical_cast<int>(const vita::value_t &v)
+{
+  using namespace vita;
+
+  switch (v.index())
+  {
+  case d_double:  return std::get<D_DOUBLE>(v);
+  case d_int:     return std::get<D_INT>(v);
+  case d_string:  return lexical_cast<int>(std::get<D_STRING>(v));
+  default:        return 0.0;
+  }
+}
+
 ///
 /// Converts a `value_t` to `std::string`.
 ///